% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanDC}
\alias{HumanDC}
\title{Human dendritic cell gene expression and meta data}
\format{
An object of class \code{list} of length 2.
}
\source{
\url{https://www.science.org/doi/full/10.1126/science.aah4573?casa_token=D7NeYw6goEAAAAAA\%3A9YM8GFN4PqkQMzXdqtX8ZJIjLnxBWuoDU1UaRRTvnUfCa-4FuWb9SNIVpUnXYbxCdK6ROkaDCdB9MYA3}
}
\usage{
HumanDC
}
\description{
A dataset human dendritic cells gene expression and corresponding meta data
}
\keyword{datasets}
