% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SCIBER_int}
\alias{SCIBER_int}
\title{Batch effect removal with SCIBER}
\usage{
SCIBER_int(
  batches_clean,
  ref_index,
  batches_meta_data,
  top_pairs_prop,
  top_genes = 50,
  n_core = parallel::detectCores(),
  combine = TRUE
)
}
\arguments{
\item{batches_clean}{A list contains all the pre-processed matrices with dimension of n_genes*n_cells.}

\item{ref_index}{The index of the reference batch in the object "batches_clean"}

\item{batches_meta_data}{A list contains the meta data for all the batches. The order should be consistent with that in "batches_clean"}

\item{top_pairs_prop}{A list of proportion of matched clusters. Default uses the 0.05 significance level for all clusters.}

\item{top_genes}{The number of marker genes used for Fisher exact test.}

\item{n_core}{Specify the number of cores otherwise use all the available cores.}

\item{combine}{TURE returns both raw and integrated batches with all batched combined. FALSE returns a list which contains all the integrated batches. The default is TRUE.}
}
\value{
A list of both raw and integrated batches with all batched combined or a list which contains all the integrated batches.
}
\description{
Batch effect removal with SCIBER
}
