% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanDC}
\alias{HumanDC}
\title{Human dendritic cell gene expression and meta data}
\format{
An object of class \code{list} of length 2.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28428369/}
}
\usage{
HumanDC
}
\description{
A dataset human dendritic cells gene expression and corresponding meta data
}
\keyword{datasets}
